Renamed repo from qcparse -> qccodec. Removed poetry in favor of hatch for a build backend and uv for project management.
