+++ .github/workflows/basic-code-quality.yaml
@@ -1,9 +1,11 @@
 name: Basic Code Quality
 
-on: [push]
+on:
+  push:
+    branches: [master]
+  pull_request:
+    branches: [master]
 
-env:
-  POETRY_VIRTUALENVS_CREATE: false
 
 jobs:
   ruff:
@@ -11,9 +13,9 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
-          python-version: "3.9"
+          python-version-file: ".python-version"
       - name: Install ruff
         run: pip install ruff
       - name: ruff
@@ -24,24 +26,24 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
-          python-version: "3.9"
-      - name: Install poetry
-        run: pip install poetry
-      - name: Install repo
-        run: poetry install --no-interaction --no-ansi
+          python-version-file: ".python-version"
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+      - name: Install the project
+        run: uv sync --all-extras --dev
       - name: mypy
-        run: mypy .
+        run: uv run mypy .
 
   detect-secrets:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
-          python-version: "3.9"
+          python-version-file: ".python-version"
       - name: Install detect secrets
         run: pip install detect-secrets
       - name: Look for secrets
