+++ .github/workflows/publish-to-pypi.yaml
@@ -13,7 +13,7 @@ jobs:
     runs-on: ubuntu-latest
     environment:
       name: pypi
-      url: https://pypi.org/project/qcparse/
+      url: https://pypi.org/project/qccodec/
     permissions:
       id-token: write # IMPORTANT: this permission is mandatory for trusted publishing
 
@@ -22,19 +22,15 @@ jobs:
         uses: actions/checkout@v2
 
       - name: Set up Python
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v5
         with:
-          python-version: 3.9
+          python-version-file: ".python-version"
 
-      - name: Install poetry
-        run: |
-          curl -sSL https://install.python-poetry.org | python3 -
-
-      - name: Install dependencies
-        run: poetry install --no-interaction --no-ansi
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
 
       - name: Build distribution 📦
-        run: poetry build
+        run: uv build
 
       - name: Publish package 📤 to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
