+++ .github/workflows/test.yaml
@@ -1,22 +1,31 @@
 name: Tests
 
-on: [push]
+on:
+  push:
+    branches: [master]
+  pull_request:
+    branches: [master]
 
 jobs:
   tests:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        name: Set up Python
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
         with:
-          python-version: "3.9"
-      - name: Install poetry
-        run: pip install poetry
+          python-version-file: ".python-version"
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+
       - name: Install repo
-        run: poetry install --no-interaction --no-ansi
+        run: uv sync --all-extras --dev
+
       - name: Run tests
         run: bash scripts/tests.sh
+
       - name: Upload coverage HTML
         uses: actions/upload-artifact@v4
         with:
