+++ .pre-commit-config.yaml
@@ -12,8 +12,8 @@ repos:
     rev: v1.4.0
     hooks:
       - id: detect-secrets
-        stages: [commit]
-        exclude: poetry.lock
+        stages: [pre-commit]
+        exclude: uv.lock
   - repo: https://github.com/charliermarsh/ruff-pre-commit
     # Ruff version.
     rev: "v0.8.3"
@@ -34,8 +34,8 @@ repos:
     hooks:
       - id: tests
         name: tests
-        stages: [push]
+        stages: [pre-push]
         language: system
-        entry: poetry run pytest
+        entry: uv run pytest
         types: [python]
         pass_filenames: false
