+++ scripts/format.sh
@@ -2,4 +2,4 @@
 
 set -x
 
-poetry run ruff check --fix .
+uv run ruff check --fix .
