+++ src/qccodec/__init__.py
@@ -1,4 +1,3 @@
-# https://github.com/python-poetry/poetry/pull/2366#issuecomment-652418094
 from importlib import metadata
 
 __version__ = metadata.version(__name__)
