+++ src/qccodec/encoders/crest.py
@@ -5,8 +5,8 @@ from typing import Any
 import tomli_w
 from qcio import CalcType, ProgramInput
 
-from qcparse.exceptions import EncoderError
-from qcparse.models import NativeInput
+from qccodec.exceptions import EncoderError
+from qccodec.models import NativeInput
 
 SUPPORTED_CALCTYPES = {
     CalcType.conformer_search,
