+++ tests/test_cli.py
@@ -1,13 +1,13 @@
 import subprocess
 
-from qcparse.codec import decode
+from qccodec.codec import decode
 
 
 def test_cli(test_data_dir):
     # Call CLI script as a subprocess
     filepath = test_data_dir / "terachem" / "water.energy.out"
     result = subprocess.run(
-        ["qcparse", "terachem", "energy", filepath], capture_output=True, text=True
+        ["qccodec", "terachem", "energy", filepath], capture_output=True, text=True
     )
     # Check the return code
     assert result.returncode == 0
